finish height setting

+++ src/urbandevelSetHeight.h
@@ -13,7 +13,7 @@ public:
     void run();
     void init();
 private:
-    int numbernearest;
+    double numbernearest;
     DM::View city;
     DM::View sb;
     DM::View sb_cent;
